Clean up translators
