+++ variant/translators/__init__.py
@@ -1,9 +1,9 @@
 """Translator package import."""
 from .translate import Translate  # noqa: F401
 from .translator import Translator  # noqa: F401
-from .polypeptide_sequence_variant_base import PolypeptideSequenceVariantBase  # noqa: F401, E501
 from .amino_acid_substitution import AminoAcidSubstitution  # noqa: F401
 from .polypeptide_truncation import PolypeptideTruncation  # noqa: F401
 from .silent_mutation import SilentMutation  # noqa: F401
 from .coding_dna_substitution import CodingDNASubstitution  # noqa: F401
 from .genomic_substitution import GenomicSubstitution  # noqa: F401
+from .coding_dna_delins import CodingDNADelins  # noqa: F401
