+++ variant/translators/genomic_substitution.py
@@ -1,10 +1,10 @@
 """Module for Genomic Substitution Translation."""
-from .dna_sequence_variant_base import DNASequenceVariantBase
+from variant.translators.translator import Translator
 from variant.schemas.classification_response_schema import ClassificationType
 from variant.schemas.token_response_schema import GenomicSubstitutionToken
 
 
-class GenomicSubstitution(DNASequenceVariantBase):
+class GenomicSubstitution(Translator):
     """The Genomic Substitution Translator class."""
 
     def can_translate(self, type: ClassificationType) -> bool:
